BLAS1 function #'scale with real or complex scalar
There are two BLAS1 functions "gsl_blas_csscal" and "gsl_blas_zdscal"
that scale complex vectors with a real scalar alpha.  In order to
distinguish a call to #'scale between these and the equivalent
functions gsl_blas_cscal and gsl_blas_zscal that take complex scalars,
it is necessary to specialize on the scalar class (float or complex).
Modified defmfun so that the pseudo-classes :component-float-type and
:element-c-type are recognized and the appropriate substitution is
made.  Compiled without errors or warnings but not tested.