Fix definition of #'euclidean-norm and defmfun :component-float-type
The function #'euclidean-norm (gsl_blas_*_nrm2) takes one vector
argument, not two, and returns the result as a return from the
function.  Handling of return declaration :component-float-type for a
method created with defmfun was wrong, because #'component-type alread
returns the C type and does not need to be converted with #'cl-ffa.