Test generation (including arrays) with save-tests
New macro #'save-tests will put list of tests onto
*all-generated-tests*.  The corresponding lisp-unit test can be
generated with #'create-test.  The macro #'generate-all-array-tests
now uses #'save-tests, so all tests defined that way will be put onto
the lists.  All tests can be written to a file with
#'write-tests-to-file, but this writes to a single file and I think I
will want to write each test to a separate file so that it is easier
to replace or augment one test without touching the others.  Two
sections, Airy and Clausen, use these definitions.