Save test forms with save-test; generate separate files with actual tests
New function #'save-test saves test forms to a special, and nothing
more.  Function #'write-test-to-file in generate-tests.lisp will write
it to a file, which ought to be done rarely.  I have started the task
of moving all test definitions to save-test, in the
special-functions.  The files generated have been checked against the
old definitions, but they have not been checked in yet.