Complex arrays initialize correctly; pool data 8-bit correct; new array tests
Complex arrays can be initialized with either reals or complexes.  If
initialized with reals, there must be an even number.  Now created
complex arrays are the right length and are not always zero in the
imaginary part.  Pool data supplied in *signed-byte-pool* is between
±127 inclusive, instead of ±255, which is required for making
(signed-byte 8).  Added some tests from functions in both.lisp to
generate-tests.lisp.