Mean and variance for all arrays
Define mean and variance functions for matrices as well as vectors.
This required a modification to #'expand-defmfun-generic and to
#'actual-array-c-type to permit the use of "both" when there is an
optional argument, and thus two C argument lists.  Defined tests for
all unweighted functions and element types.  Because there appears to
be a bug in the GSL library in computing single floats, no tests have
yet been defined for these functions.