Tests for BLAS2 and BLAS3
Added tests for all array element types for BLAS2 and BLAS3.  Those
functions that take complex scalars as arguments will not work due to
CFFI's lack of support for passing structs directly.  The rank n
update functions in BLAS2 gives an "Matrix, vector lengths are not
conformant" error (EBADLEN), which needs to be tracked down.
Therefore, none of these tests are defined in either file.  Fixed
arguments to C function in #'matrix-product-symmetric and
#'inverse-matrix-product in BLAS3.