Delete early experiment with foreign-friendly arrays
In early experimentation with foreign-friendly arrays, I developed a
way to directly use the foreign-friendly arrays in GSL code using raw
C arrays.  This was successful.  Subsequently, I rewrote the gsl-data
objects to have foreign-friendly arrays underlying them, whether they
create raw C arrays, GSL arrays, or both, so the early experiment is
obsolete and is deleted in this commit.