Analysis of port progress; improve documentation
Started in file init/analysis.lisp some definitions that will aid in
gauging how much of GSL GSLL covers.  It will not normally be loaded
(and depends on port system which is not required for GSLL) because it
is mainly used by developers.  It is not yet finished.  Hash size for
*gsl-symbol-equivalence* increased to reflect actual number of symbols
present when GSLL is loaded.  Improved documentation.