Native arrays do not need wrapper
Native arrays are now completely dynamic, i.e., they are garbarge
collected and there is no need for any kind of wrapper, because the
with-pinned-objects is wrapped at the defmfun level.  Most tests pass
on SBCL, but there are still a number that fail.  The GSL parts are
always present; there is potentially unnecessary allocation of all
GSL-required structs at creation time, which might merit some
investigation for a more just-in-time creation procedure.  Potential
problems on non-native implementations completely uninvestigated.