Combinations are garbage-collected; restore floating-point and mathematical
Redefine #'alloc-gsl-struct as a generic function and define a method
specifically for combinations.  Define #'native-pointer in
defmfun.lisp with a new function #'native-pointer-protect made from
its body left in data.lisp, so that now floating-point and
mathematical can be compiled and loaded as defined in the asd file.
Use :pointer for foreign argument types instead of gsl-vector-c.