GSL array making with make-array* and #m
GSL arrays are garbage-collected objects, and so can be made like any
other lisp object; they do not need to be make inside a wrapper form.
#m is a macro which makes them; it expands to a #'make-array* call.
Examples and usage ported to this new form.  Compiled only; not tested
yet.  No effort to port to non-native implementation yet; tests as
expanded in the tests/ directory are not ported yet.