Make FFA at initialize-instance time for gsl-data; restore permutations and combinations
Instead of making the ffa in make-array*, we now make it in the
initialize-instance :after method for gsl-data.  This means that
subclasses (permutations, combinations) will automatically make the
arrays.  Rewrote make-array* so that it is just a wrapper for
make-instance, computing the dimensions from the specified initial
parameters.  Permutations and combinations are now made with make-*
functions.