Changes to defmfun to permit method qualifiers and &key arguments
Add :qualifier argument to defmfun, used when methods are defined to
allow qualifiers such as :after or :before.  Permit &key arguments.
Check the :after argument to defmfun for usage of argument variables
to prevent them from being declared ignored.  Note:  This uses
stupid-code-walk-find-variables, and is, well, stupid.  These changes
are all designed to permit the definition of #'initialize-instance and
#'reinitialize-instance methods for GSL objects.