New class mobject, new hierarchy, clean up files
Redefine 'mobject from 'gsl-object.  The class 'gsl-data in data.lisp
is now a subclass of it, and does not explicitly have the mpointer
slot, because mobject does.  Adapted #'defmobject to work with objects
that have no "setter" (initialization).  This expands correctly for
basis-splines (no setter) and hankel (setter).  The basis-splines
defintions rely on the new mobjects.