GSL object makers with mixed arguments
GSL objects made with defmobject and that have use at least one
argument in some way for both initialization and reinitialization, or
for using functions of those arguments, are accomodated with an
optional argument 'arglists-function to defmobject.  This should be a
function that produces three arglists: for the maker, the
initialization method, and the reinitialization method.  The order of
arguments generated for other cases is now preserved with :from-end in
#'variables-used-in-c-arguments.  Examples for interpolation and
acceleration (lookup), still commented out.