Set mpointer vlaue, show relevant .h file; non-int return for reinitialize-instance
In defmobject, set the mpointer value if #'allocate is being called.
For each file that defines a defmobject, I put a comment at the top
showing the relevant GSL header file.  For generators and quasi, the
(re)initializing function returns void and not an :int, so that is
changed in the defmobject definition, which now permits a list of two
elements in initialize-suffix, with the second element indicating the
:c-return.