GSL arrays created by GSL functions
Made the generic function #'contents-from-pointer that takes a pointer
to a GSL struct and produces a list of the contents suitable for the
:initial-contents argument to #'make-marray.  Added an argument
:from-pointer to #'make-marray that then creates a new marray with the
same contents.  Some functions in solve-minimize-fit return a pointer
to a GSL vector with double-floats.  Copying in this way is necessary
because GSL has malloced the data in a non-CL place.  Sometimes, this
results in the copying of data whose only further use will be to
extract a pointer to pass to another GSL function.  Added a method
to #'mpointer that permits directly passing a pointer, but this isn't
useful for arrays because there are many element types so the
functions have been defined as methods.