Redesign polynomial-solve, invert matrix
The function #'polynomial-solve is now directly defined with defmfun
with a single required argument and two optional arguments: the answer
vector-complex-double-float and the workspace, both of which default
to the right thing.  The returned is now the this marray instead of a
list.  Simplify the definition of #'invert-matrix and return the
matrix-double-float instead of the CL array.  All tests pass SBCL and
CCL with trace.