Fully functional on CCL
After exiting the foreign call, defmfun now inserts a form to set
C-invalid to nil for outputs, as well as CL-invalid to T.
Conditionalize the definition of native-pointer to native, because
it's only used on native implementations and its definition in CCL was
causing a spurious warning.  Removed defmfun expansion examples
because there are plenty of examples available now.  From scratch,
gsll and gsll-tests compiles and runs cleanly on CCL with no warnings
or errors, and all 1208 tests pass.