Documentation current
Rewrote the documentation to reflect the current state of the
software, as a single HTML file index.html, and used a new css style
cribbed from iterate.  There are two text files added: status.text
that is a summary of test pass/fail on various implementations and
platforms, and missing-features.text that documents features of GSL
that aren't included.