Elementwise array operations named "elt"
GSL functions that operate from one corresponding array elements and
then overwrite the first array, as well as those that operate on each
element of an array with a scalar, are now named starting with "elt",
as in "elt+"; formerely they started with "m".  This is to reinforce
the idea that they are not generally a mathematical solution to
algebraic problems, just elementwise operations.  The scalar
operations (addition, multiplication) are methods of elt+ and elt* as
well. Test results 64 bit:
SBCL: TOTAL: 1223 assertions passed, 5 failed, 0 execution errors.
CCL: TOTAL: 1228 assertions passed, 0 failed, 0 execution errors.