Default optional/key argument values for arrays
When used as the default optional/key argument values for arrays, the
value symbol 'element-type is replaced with the actual element type
for the method.  This replaces the :make-marray form.  When used as
the default optional/key argument values for arrays, the form
(eltcase element-type form element-type form ...)
will be replaced by the appropriate form.  This replaces the direct
appending of the (element-type form element-type form) list.
SBCL same tests pass.