Redefine defmcallback to provide and accept scalars from CL functions
New definition for defmcallback will now convert C array pointers
passed to the function by GSL into individual scalar arguments of the
CL function, and will take multiple value returns and assign as
elements of the return C arrays.  Applied to def-ode-functions and the
van der Pol oscillator example.  Redefined #'maref-function-picker as
a macro from a function, eliminating the need for an eval-when form.