Minimization will take scalar functions, added tests, optional printing
The minimization functions may now take and return scalars or marrays.
The examples take optional arguments with the solver method and
whether to print the steps.  They are now part of the #'examples and
test suite.  Use maref instead of cl-array in
#'roots-multi-example-no-derivative.  Add optional solver method to
roots one and roots multi, and add tests for all solvers.