Polynomial uses #'evaluate
Polynomial evaluation is now done with a method of #'evaluate.  We
can't do complex evaluation because the return value would be complex,
and there is no way to handle that in CFFI yet.  Evaluation with
divided differences is a keyword switched part of the same method; it
works but gives an puzzling style warning in SBCL about the keyword.
The use of an marray in an optional/key argument necessitated a change
to defmfun that redefined #'body-expand from #'body-no-optional-arg,
made #'body-no-optional-arg a new function that wraps necessary
array-handling forms around the expanded unswitched body form from
body-expand.