Split up definition of defmobject, add definition of defun*
The definition of defmobject has become too large so the parts that
are responsible for construction of the various functions are broken
out as functions:
  make-initialize-instance
  make-reinitialize-instance
  mobject-maker

The definition of callbacks for GSL will be changed so that the
callback function (defmcallback) is defined at the time the function
is defined, and the callback GSL structure is made (defcbstruct) at
the time the corresponding defmobject is made, along with a freeing
function to add to the finalizers.  (For callbacks used in functions
without mobjects, the struct will be allocated before and then freed
after the call.)  As a first step toward this organization, defun* is
defined to handle in one form the definition of a CL function and of a
callback.