Hacked changes define cbstruct in mobject maker
Have the mobject maker build the cbstruct.  This is now hacked in and
works with roots-multi in #'roots-multi-example-no-derivative, but
with #'make-multi-dimensional-root-solver-f having a hardwired
dimension of 2.  No clear way to communicate the dimension(s) from the
other part of the defmobject form.  Started to think/write a subclass
'callback-included of 'mobject, but this may be overkill as all
variables except 'dimensions are exclusive to the cbstruct maker
itself so don't really need to be class slots.