Define callbacks with #'make-callbacks form at top level
Defined a new macro #'make-callbacks that is invoked at top level.  It
is given the mobject class name, the function name(s), dimension(s)
and whether it should assume scalar input/output.  It will define the
appropriate callbacks for use in the mobject.  This eliminates the
need for defun* and should work with all CL implementations.  Tested
on both roots-multi only.