Build 'one-dimensional-root-solver-f on 'callbacks-included
Now define one-dimensional-root-solver-f on 'callbacks-included.  This
needs some modification to mobject to add the 'functions slot argument
and remove 'dimensions from the cbstruct when it is not in the
allocation arguments.  The :functions argument must be given as a
list, even though there's only one function.  The example
(roots-one-example-no-derivative) works.