One-dimensional root solver ported to callbacks-included
The one-dimensional root solvers in roots-one.lisp now works in the
callbacks-included object, including both examples, with and without
derivatives.  The no-derivative case has been modified to
automatically listify the function name given.  For callbacks-included
classes that do not define 'dimensions (because they are
one-dimensional), there is a new macro def-ci-subclass-1d that makes
'dimensions slot a class slot with '(1).  Defined a print-object
method for callbacks-included classes.