Make mobjects with :singular arguments
Some parameters to the make-* function for mobjects are taken as lists
(like 'functions, 'dimensions) but for some mobjects it only makes sense
to supply one.  In these cases, the :singular argument to defmobject
indicates that that argument is to be put in a list before being
passed to the next function.