Nonlinear least squares ported to callbacks-included; 'dimension-names
Nonlinear least squares now works using callbacks-included.  This
necessitated naming the components of dimensions individually 
(number-of-observations number-of-parameters).  This is done with
added slot 'dimension-names to the callbacks-included class and an
argument of the same name to #'def-ci-subclass which defaults to
'(dimensions).  Because of the default, the other defmobject usage
need not be changed.