Consistent apply-step, apply-evolution function names and arguments
The arguments have been changed in apply-step, apply-evolution so that
they match for the first three arguments.  The function name
apply-step was changed so the two are similar.  The macro
with-ode-integration defines a local function #'next-step instead of a
symbol macro 'make-next-step because symbols in the body of a simple
loop violate the CL standard.  Some documentation has been updated
with the contribution of Mirko Vikovic.