Make 'single-function "class" for (def-)make-callbacks
Many callback users need just a single function taking and returning
scalars.  These are all defined the same way.  So I have defined in
callbacks.lisp a def-make-callbacks for 'single-function.  This works
for chebyshev, minimization-one, and roots-one, and it should be
applicable to callbacks sent directly to functions without being
stored in mobjects.