Direct use of callbacks in functions
Simple function passing to functions which don't use mobjects is
possible with some modifications to defmfun which treat the C argument
'callback and the CL argument 'function specially, generating a
callback structure based on 'gsl-function and filling it
appropriately.  Ported numerical-differentiation and
numerical-integration.  Also simplified calls in numerical-integration
by making the workspace argument optional and at the end of the
arglist; it defaults to the right size based on 'limit.  Changed one
of the tests for numerical-integration because the limit test doesn't
make sense with the workspace automatically allocated.