GSL features added in version 1.9
Added definitions for GSL features that were added in version 1.9, and
one from version 1.10: mminusp, mplusp, non-negative-p,
+knuthran2002+, +vector-bfgs2+.  Macros defmpar now has key instead of
optional args.  It and def-rng-type take a gsl-version argument; any
symbols that are not defined in the user's version of GSL will be
bound to an instance of 'obsolete-gsl-version but it is not signalled.