GSL features added in version 1.11; cholesky-solve and linear fit unification
Added +halton+, +reverse-halton+ to random/quasi.lisp. 

Added #'multi-linear-residuals in
solve-minimize-fit/linear-least-squares.lisp.  Changed +sobol+,
+niederreiter2+ from *..* apparently overlooked before.

#'cholesky-solve now written so that if x-spec = T, a marray is
created and the solution put in it and returned.  If it is NIL, the
solution will replace b.  This required a modification to
expand-defmfun-generic to remove &aux args from the defgeneric
arglist.

solve-minimize-fit/linear-least-squares.lisp
Specify the weight as an optional argument, unifying the weighted and
unweighted functions:
linear-fit, multiplier-fit, linear-mfit, linear-mfit-svd.
This necessitated a change in the order of the arguments.
Changed call in basis-splines.lisp.