Optional arguments for linear algebra functions; unification of solve and svx
Most linear algebra functions now have optional arguments which create
workspace and answer arrays.  All solver function that have _solve
(separate solution vector) and _svx (solve in-place) forms have been
unified into a single CL function.  In qrpt functoins where signum is
computed, it is no longer taken as input but allocated internally and
return as one of the multiple values.  This necessitated a change in
body-expand that will now apply the cl-convert-form conversion to
returned variables that are part of a multiple value list for
:return.