Scalar functions passed as callback arguments
A dynamic (special) variable whose name is gensymmed is used to bind
the function passed by the user as the argument to the GSLL function.
The defmfun that defines this function also sets the callback
structure to another gensymmed name, and it calls a function that
funcalls the dynamic variable.  This has been implemented for
integration-QNG and works.