In defmfun, :callbacks argument extended; formalize reading
The :callbacks argument to defmfun is now has at least four elements,
the first two specifying the callback argument name and the structure
name, and the remaining pairs specify the slot name and name of the
variable bound to the function.  Functions defined for extracting
these pieces: callback-carg, callback-structure-type,
callback-slots-fns.  Even though in defmfun it is not necessary to
define more than one callback function, it is possibile to define
more.  This has been tested on integration-QNG only.