Dynamic callback functions in defmfun for array arguments
Extended the dynamic binding of callback functions to include a flag
on whether to use scalars, and the dimension of the arguments, so that
callback functions that have more than one scalar input and are
defined in a defmfun will work.  This includes exactly Monte Carlo
functions.  Scalar argument functions in numerical differentiation and
numerical integration still work.