Separate :callback-dynamic from :callbacks arguments
The :callback argument to defmfun has now been split into two pieces,
a set of values that will be specified at run time, which is in
:callback-dynamic, and the values needed to build the callback and
the callback structure, which remains in :callback.  The contents are
explained in callback.lisp.  For mobject callbacks, the :callback part
will be in the defmobject, and the :callback-dynamic will be in the
defmfuns that use the mobject.  Currently, the dimensions template in
:callback is not read, just the first dimension in :callback-dynamic
is used.  This needs to be fixed.  The numerical-integration,
numerical-differentiation, and monte-carlo examples work.