Dynamic dimensions for callbacks
Dimensions that have been placed into the dynamic-variable are now
read and used for array input, as for example in Monte Carlo.
Functions that use callbacks with scalar input and output (as for
example in numerical-integration and numerical-differentiation) are
now expanded directly with a funcall instead of call-maybe-scalar.
Modifications and new definitions to set arrays partly written and
untested, await changes to defmobject where callbacks setting array
values exist.