Use tmh's lisp-unit definitions instead of old
Incorporated tmh's lisp-unit definitions he posted to
http://paste.lisp.org/display/76579 in preference to the former
definitions in additional-definitions.lisp.  He took some of those
ideas and incorporated them with his own.  This produces some
mysterious errors not seen before when run on both SBCL and CCL
(different errors for each).

The plan is to spin off lisp-unit as a separate system which will just
be loaded by gsll-tests.