New functions to parse :callbacks argument
New functions 
  parse-callback-static
  parse-callback-fnspec
  parse-callback-argspec
work on the :callbacks argument.  The first one parses the whole
argument.  The second one parses a particular function specification
from :callbacks.  The third one parses a particular argument of a
particular function.  When given 'dimensions, this function supplies
all the dimensions, and #'defmcallback expands to make a list of the
arguments.  Then call-maybe-scalar passes these on to array-to-list,
which currently only handles a vector.  These are now used in
#'make-defmcallbacks and #'defmcallbacks.  Both
monte-carlo-integrate-plain and integration-QNG compile and run
correctly.  I have eliminated the old defmcallback and supporting
definitions to clean up the file.