Dynamic callback defined in mobject, works for chebyshev
The class 'callback-included has been simplified by replacing several
obsolete slots with 'callbacks and 'callback-dynamic.  The :callbacks
static information is saved to a hash table so that it may be
retrieved when the defmfun is expanded.  The function
make-cbstruct-object to make the callback structure
as an &aux variable in reinitialize-instance :after form expanded in
defmobject is now a form expansion executed at macroexpand time.
The argument :callback-object has been added to defmfun; this will
result in the binding of the dynfn dynamic variable on expansion.

The callback arguments for chebyshev definitions have been updated and
the example works.