Better support for array returns (prototype roots-one), ordering of load
In roots-one.lisp, added arguments for dynamic callbacks.  This is not
complete; 
 (roots-one-example-derivative +newton-fdfsolver+ nil)
causes both SBCL and CCL to crash to low-level debuggers.
These changes included using a new function list-to-arrays in
call-maybe-scalar so that multiple arrays can be set. 
New file callback-compile-defs has definitions that are needed at
compile time of later files.  Added a call record-callbacks-for-class
in defmobject, in addition to the expansion in the form, because the
hash table *callbacks-for-classes* needs to be updated with the class
being defined for the macro make-cbstruct-object to work.  Added
comments above defmobject to explain some of the arguments.