Generalize elt-, elt/; fix docstring; export 'dim0, 'dim1
Add methods for #'elt- and #'elt/ that take scalars for the second
argument.  Fix the docstring for (setf row) which had "column" instead
of "row".  Export the symbols 'dim0 and 'dim1 because they are useful
for users; use them in #'contents-from-pointer instead of previous
dim1 and dim2 (et. sim. with size0, size1).