Fixed arguments to linear least squares functions
In both linear-mfit-nosvd and linear-mfit-svd, the arguments
were in the wrong order, and the there was no #'mpointer call for
covariance.  There was a 'tolerance argument in the C arglist for
non-svd function, which should not have been there.  Also, 'weight
needed to be declared as an input.  This last was the cause of two
test failures in non-native implementations.  Now
CCL64: TOTAL: 1454 assertions passed, 2 failed, 0 execution errors.
SBCL64:TOTAL: 1451 assertions passed, 5 failed, 0 execution errors.