Dynamic callbacks for minimization-multi
Argument 'scalarsp added at end of optional arguments to mobject
maker; it is then passed to reinitialize-instance.  Even though it is
not specified as an argument there, it is accepted to set the slot of
that name.  Ported minimization-multi to dynamic callbacks; examples
all work.  Linear least squares re-enabled in gsll.asd; no changes.