New errno code interpretation and gsl-condition definitions
Make a defconstant on earmuffed symbols of cffi:defcenum gsl-errorno,
e.g. +continue+, +failure+, +success+, etc., and use those in
define-gsl-condition forms in condition.lisp.  Remove the symbol macro
definition of 'success.  Define a condition 'unspecified-errno in case
an error code is processed that isn't in the list.  Redefine
#'success-failure to true if either +success+ or +continue+, from
non-negative.  

In roots-multi, fixed error in multiroot-test-residual return for, and
changed roots-multi-example-no-derivative to insure the loop is
executed at least once so that fnval and argval are always bound
before exiting.