Start port new spec of callback arguments, make-funcallable, ODE
New specification of callback arguments
parse-callback-argspec takes 
  io, element-type, array-type, dimensions
Not at all implemented anywhere in the code.
Made file funcallable.lisp out of the old defmcallback, designed to
create a lambda for compilation that will be funcalled by the (new)
defmcallback body.  This still uses the old specification for function
arguments.  Started porting ODE callbacks, but this has been suspended
while the first two changes are being completed.  None of this works
in any way.