Incorporation of make-funcallable into callback call chain
The callback call chain (for mobjects at the moment) now calls
make-funcallable when the mobject is made.  The :callbacks format is
slightly different in that :slug is required.  The derivative-free
minimization-multi example works, but requires a double quote on the
function name, which should be fixed.  The runtime-called function
#'call-maybe-scalar and its supporting definitions have been removed,
in anticipation of all callbacks using make-funcallable.