Most mobject-based callbacks ported to funcallables
All mobject-based callbacks now use the funcallable mechanism, except
ODE, which has been ported but fails to compile.  All other tests
pass.  In the event of some or all arguments are scalars and not
arrays, #'array-element-refs will return just the argument symbol.  To
avoid double quoting the function name when calling the object maker,
let bind 'function-designator in #'make-funcallable-form which does
the quoting.