Number symbols sequentially; intern specials for defmfun
Created file init/utility.lisp, which now includes
make-symbol-cardinal and make-symbol-cardinals.  These functions in
turn will now take a list for name, and join the symbols together with
"-", and will optionally intern in a package.  In #'callback-args in
callback.lisp, the args are now numbered sequentionally (still
uninterned, but no longer gensymmed).  In #'expand-defmfun-wrap in
defmfun.lisp, the symbols for specials "dynfn" and "cbfn" that are
made are now interned, numbered sequentially, and named after the
function in which they occur.  This doesn't affect the functionality
of the functions, it is just for readability of the macroexpansion.
Expected tests pass.