Use FSBV; complex-types.lisp work
Use Foreign Structures By Value (FSBV) to handle passing and returning
complex scalars defined by one of the GSL structs.  This works
correctly for the four functions in complex.lisp, but there is
compiler error in SBCL on set-all: "cannot use values types here",
though the expansion looks correct.  Does not work on
solve-quadratic-complex and solve-cubic-complex in polynomial.lisp
because defmfun takes the structure type specified and converts it to
:pointer automatically.  Not tried on functions that return complex
scalars yet.