New macro foreign-funcall-indirect
Move body-expand and the supporting cast to a new file
body-expand.lisp, and introduce foreign-funcall-indirect which will
use with-foreign-objects to allocate all arguments at additional level
of indirection and make the appropriate assignments.  This appears to
expand correctly except for values that are not symbols,
e.g. (mpointer object) used in set-all.  I will need a new syntax for
specifying structures that are passed by pointer but which need to
have the structure type stated, e.g. solve-cubic-complex, because for
now it defmfun will expand into the foreign-funcall-indirect.